Reduce width of millis_t to save flash space

+++ firmware/include/millis.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <stdint.h>
 
-typedef uint32_t millis_t;
+typedef uint16_t millis_t;
 
 millis_t millis();
